Set up basic security test code
Set up some test authentication code in AppUser model, also updated some files to add proper tags. Changed model ID data types to "long" instead of "int"

+++ backend/src/main/java/com/skillstorm/backend/Repositories/RoomTypeRepository.java
@@ -3,6 +3,6 @@ package com.skillstorm.backend.Repositories;
 import com.skillstorm.backend.Models.RoomType;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface RoomTypeRepository extends MongoRepository<RoomType, Integer> {
+public interface RoomTypeRepository extends MongoRepository<RoomType, Long> {
     // Custom query methods if needed
 }
