+++ backend/src/main/java/com/skillstorm/backend/Repositories/TransactionRepository.java
@@ -3,6 +3,6 @@ package com.skillstorm.backend.Repositories;
 import com.skillstorm.backend.Models.Transaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface TransactionRepository extends MongoRepository<Transaction, Integer> {
+public interface TransactionRepository extends MongoRepository<Transaction, Long> {
     // Custom query methods if needed
 }
